Исправления без стилей
