+++ lib/screens/Registr_Screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:rickmorty/login_screen.dart';
+import 'package:rickmorty/screens/login_screen.dart';
 
 class RegScreen extends StatelessWidget {
   const RegScreen({super.key});
@@ -7,26 +7,23 @@ class RegScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: const Color(0xff0B1E2D), // Исправленный цвет
+      backgroundColor: const Color(0xff0B1E2D), 
       body: SingleChildScrollView(
         child: Padding(
           padding: const EdgeInsets.symmetric(horizontal: 28), // Общий отступ слева и справа
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start, // Выравнивание элементов слева
             children: [
-             // const SizedBox(height: 26), // Отступ сверху вместо margin
-              Container(
-                margin: EdgeInsets.only(top: 11,bottom: 28),
-                child: SafeArea(child:
-              GestureDetector(
-                onTap: () {
-                  Navigator.pop(context);
-                },
-                child: const ImageIcon(AssetImage('assets/images/back.png'),color: Colors.white,),
-              ),),),
+              const SizedBox(height: 54), // Отступ сверху вместо margin
 
-              //const SizedBox(height: 28), // Отступ после стрелки
+                 GestureDetector(
+              onTap: () {
+                Navigator.pop(context);
+              },
+              child: const ImageIcon(AssetImage('assets/images/back.png'),color: Colors.white,),
+                            ),
 
+              const SizedBox(height: 28), // Отступ после стрелки
               const Text(
                 "Создать аккаунт",
                 style: TextStyle(
@@ -37,7 +34,6 @@ class RegScreen extends StatelessWidget {
               ),
 
               const SizedBox(height: 40), // Отступ перед полем "Имя"
-
               const Text(
                 'Имя',
                 style: TextStyle(
@@ -103,7 +99,7 @@ class RegScreen extends StatelessWidget {
               ),
               const SizedBox(height: 36,),
               Divider(
-                color: Color(0xfff152a3a).withOpacity(0.5),
+                color: const Color(0xfff152a3a).withOpacity(0.5),
                 thickness: 2,
               ),
               const SizedBox(height: 36,),
@@ -118,8 +114,8 @@ class RegScreen extends StatelessWidget {
               const SizedBox(height: 8),
               TextField(
                 decoration: InputDecoration(
-                  prefixIcon: const Icon(Icons.person, color: Colors.grey),
                   hintText: 'Логин',
+                  prefixIcon: const ImageIcon(AssetImage('assets/images/User.png'),color: Color(0xfff5B6975),),
                   filled: true,
                   fillColor: const Color(0xfff152A3A),
                   border: OutlineInputBorder(
@@ -127,7 +123,6 @@ class RegScreen extends StatelessWidget {
                     borderSide: BorderSide.none,
                   ),
                 ),
-
               ),
               const SizedBox(height: 10,),
               const Text(
@@ -142,7 +137,7 @@ class RegScreen extends StatelessWidget {
               TextField(
                 obscureText: false,
                 decoration: InputDecoration(
-                  prefixIcon: const Icon(Icons.password_outlined, color: Colors.grey),
+                  prefixIcon: const ImageIcon(AssetImage('assets/images/password.png'),color: Color(0xfff5B6975),),
                   hintText: 'Пароль',
                   filled: true,
                   fillColor: const Color(0xfff152A3A),
@@ -153,37 +148,28 @@ class RegScreen extends StatelessWidget {
                 ),
               ),
               const SizedBox(height: 63,),
-              Container(
-                child: SizedBox(
-                  width: double.infinity,
-                  child: ElevatedButton(
-                    onPressed: () {
-                      Navigator.pushReplacement(
-                        context,
-                        MaterialPageRoute(builder: (context) => const LoginScreen()),
-                      );
-                    },
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: const Color(0xFF41B9E5),
-                      padding: const EdgeInsets.symmetric(vertical: 14),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(10),
-                      ),
-                    ),
-                    child: const Text(
-                      "Создать",
-                      style: TextStyle(fontSize: 16, color: Colors.white),
+              SizedBox(
+                width: double.infinity,
+                child: ElevatedButton(
+                  onPressed: () {
+                    Navigator.pushReplacement(
+                      context,
+                      MaterialPageRoute(builder: (context) => const LoginScreen()),
+                    );
+                  },
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: const Color(0xFF41B9E5),
+                    padding: const EdgeInsets.symmetric(vertical: 14),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(10),
                     ),
                   ),
+                  child: const Text(
+                    "Создать",
+                    style: TextStyle(fontSize: 16, color: Colors.white),
+                  ),
                 ),
-
               ),
-
-
-
-
-
-
             ],
           ),
         ),
